+++ components/react/src/index.ts
@@ -1,1 +1,0 @@
-export { HighlightComponent, TextSearch } from './TextSearchHighlight'